Introduce file-based user configuration
This PR addresses https://github.com/jotaen/klog/discussions/158.

klog now looks for a config file at `~/.klog/config.yml`, where users
are able to specify some preferences. For now, that is:

- `default_rounding`: the value for the `--round` flag, unless
specified, e.g. in `klog start` or `klog stop`.
- `default_should_total`: the default for the `--should-total` flag,
unless specified, that is used whenever klog creates a new record, e.g.
in `klog create`, but also in `klog track` or `klog start`, as these
might create new records on the fly.

More configuration parameters may follow later, now that the groundwork
is set up.

When the config file is absent, everything behaves as before, so it’s
strictly optional right now.

What’s also new is that klog will respect the `$KLOG_FOLDER_LOCATION` or
`$XDG_CONFIG_HOME` environment variables, if set. In this case, **and**
when users have bookmarks, then this will be a breaking change. The fix
is for them to manually move over their `bookmarks.json` file to the new
klog folder location. I’ll take note in the Changelog about this.

The `klog config` command is hidden from the help output for now, but it
can be invoked nevertheless on the nightly build (when compiling klog
from the `main` branch). I’ll do some more testing, refactoring, and
enhancements in the next time, and plan to include this with the next
release – probably within the next few weeks.