Clean up install instructions
The structure of the install instructions are a bit messy, which this PR
aims to straighten up.

The idea is this:

- The [documentation website](https://klog.jotaen.net#get-klog) is the
canonical place for all install options and instructions.
- The `INSTALL.md` file that’s contained in the `.zip` file of the
Github releases should only contain the relevant parts for the manual
installations. (The user has already downloaded the zip file, and thus
decided that they want to install manually, after all.) It’s not meant
as comprehensive reference for install options.