Add `--resume` flag for `klog start`
Resolves https://github.com/jotaen/klog/issues/258.

The `--resume` / `-R` flag takes over the last entry summary when
starting a new entry. If there is no previous entry, or if the previous
entry summary is empty, it doesn’t create a summary for the new entry.
The `--resume` flag cannot be used together with the `--summary` flag.

Note, the `-r` shorthand flag was already reserved by `--round`.