Add command `klog switch`
Resolves https://github.com/jotaen/klog/issues/268.

`klog switch` stops an ongoing activity (open range) and starts a new
one at the same time. It’s basically a convenience shortcut for
separately issuing `klog stop` and `klog start` – with a few subtle
differences, though:

- The command is atomic, so it only writes to the file once. The end
time of the previously open range and the start time of the newly opened
range are guaranteed to be the same.
- The command doesn’t have fallback-behaviour. It neither falls back to
the previous date like `klog stop` does, nor would it create a new
record like `klog start` would.
- A summary for the new entry can be specified via the `-s`/`--summary`
flag. Appending to the summary for the previously open range is not
possible.

The command deliberately fails if there is not open range in the record.